requirements.txt
fastapi==0.111.0

langchain==0.3.13
langgraph==0.2.60 
langchain-core==0.3.28
langsmith==0.1.143 

langchain-openai==0.2.1
langchain-google-genai==2.0.0
langchain_anthropic==0.2.1
langchain-groq==0.2.0
langchain-community==0.3.1
langchain-ollama==0.2.0 
langchain-nvidia-ai-endpoints==0.3.7 
langchain-deepseek-official==0.1.0 

langchain-text-splitters==0.3.4

pydantic==2.9.1
python-dotenv==1.0.1
uvicorn==0.29.0
PyYAML==6.0.1
upstash-redis==1.2.0 


passlib==1.7.4 
PyJWT==2.9.0 

SQLAlchemy==2.0.34 
psycopg2==2.9.9 

beautifulsoup4==4.12.3
langchain-google-community==2.0.0 

cohere==5.11.0 

google-api-python-client==2.149.0 
tavily-python==0.5.0 
duckduckgo-search==6.3.2 
exa-py==1.4.0
wikipedia==1.4.0 

pypdf==5.1.0 
langchain-unstructured==0.1.5 
firecrawl-py==1.5.0 

langchain-milvus==0.1.7 
langchain-cohere==0.3.1 
pymilvus==2.4.9

python-multipart==0.0.19 

# docstring==0.1.2.4 

# transformers==4.42.0 
# fastapi-cors==0.0.6